fix(ci): SHUT_RDWR is Int on Linux in hop test server
shutdown() wants Int32; Darwin's SHUT_RDWR is already Int32, Linux's is Int.

+++ Tests/BarkVisorTests/SSRFProtectionTests.swift
@@ -479,7 +479,7 @@ private final class SSRFHopHTTPServer: @unchecked Sendable {
     }
 
     func stop() {
-        shutdown(fd, SHUT_RDWR)
+        shutdown(fd, Int32(SHUT_RDWR))
         close(fd)
     }
